refactor into better folder structure, still more to go
