+++ src/config.ts
@@ -10,7 +10,7 @@ import * as os from 'os';
 import * as telemetry from '@cmt/telemetry';
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
-import { CppDebugConfiguration } from '@cmt/debugger';
+import { CppDebugConfiguration } from '@cmt/debug/debugger';
 import { Environment } from '@cmt/environmentVariables';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
