+++ src/debug/cmakeDebugger/cmakeDebuggerTelemetry.ts
