+++ src/extension.ts
@@ -31,7 +31,7 @@ import rollbar from '@cmt/rollbar';
 import { StateManager } from './state';
 import { cmakeTaskProvider, CMakeTaskProvider } from '@cmt/cmakeTaskProvider';
 import * as telemetry from '@cmt/telemetry';
-import { ProjectOutline, ProjectNode, TargetNode, SourceFileNode, WorkspaceFolderNode } from '@cmt/projectOutline/projectOutline';
+import { ProjectOutline, ProjectNode, TargetNode, SourceFileNode, WorkspaceFolderNode } from '@cmt/ui/projectOutline/projectOutline';
 import * as util from '@cmt/util';
 import { ProgressHandle, DummyDisposable, reportProgress, runCommand } from '@cmt/util';
 import { DEFAULT_VARIANTS } from '@cmt/kits/variant';
@@ -41,13 +41,13 @@ import { CMakeDriver, CMakePreconditionProblems } from './drivers/cmakeDriver';
 import { platform } from 'os';
 import { CMakeToolsApiImpl } from '@cmt/api';
 import { DirectoryContext } from '@cmt/workspace';
-import { ProjectStatus } from '@cmt/projectStatus';
-import { PinnedCommands } from '@cmt/pinnedCommands';
+import { ProjectStatus } from '@cmt/ui/projectStatus';
+import { PinnedCommands } from '@cmt/ui/pinnedCommands';
 import { StatusBar } from '@cmt/status';
-import { DebugAdapterNamedPipeServerDescriptorFactory } from '@cmt/debug/debugAdapterNamedPipeServerDescriptorFactory';
-import { getCMakeExecutableInformation } from '@cmt/cmake/cmakeExecutable';
-import { DebuggerInformation, getDebuggerPipeName } from '@cmt/debug/debuggerConfigureDriver';
-import { DebugConfigurationProvider, DynamicDebugConfigurationProvider } from '@cmt/debug/debugConfigurationProvider';
+import { DebugAdapterNamedPipeServerDescriptorFactory } from '@cmt/debug/cmakeDebugger/debugAdapterNamedPipeServerDescriptorFactory';
+import { getCMakeExecutableInformation } from '@cmt/cmakeExecutable';
+import { DebuggerInformation, getDebuggerPipeName } from '@cmt/debug/cmakeDebugger/debuggerConfigureDriver';
+import { DebugConfigurationProvider, DynamicDebugConfigurationProvider } from '@cmt/debug/cmakeDebugger/debugConfigurationProvider';
 import { deIntegrateTestExplorer } from "@cmt/ctest";
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
