+++ src/ui/projectOutline/targetsViewCodeModel.ts
