+++ test/unit-tests/driver/driver-test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-unused-expressions */
-import { CMakeExecutable, getCMakeExecutableInformation } from '@cmt/cmake/cmakeExecutable';
+import { CMakeExecutable, getCMakeExecutableInformation } from '@cmt/cmakeExecutable';
 import { ConfigureTrigger } from '@cmt/cmakeProject';
 import { ConfigurationReader } from '@cmt/config';
 import * as chai from 'chai';
